Use addOpens to get around unnamed caller warnings
Core JDK classes always appear open to the unnamed package, so
we can't simply use isOpen here. Instead, we combine that with a
check for the unnamed package, and use addOpens to make it "really
open" to eliminate the warning that would normally result.

This also modifies the trySetAccessible logic to construct no new
objects (e.g. Module9 wrappers) by going straight to the Java 9
APIs after a version guard.